requirements.txt
tensorflow
google-cloud-storage
Pillow
librosa
scikit-learn
fastapi
uvicorn